Update README.md
I tested with in develop and build, and I didn't see any query errors. I'm guessing that was from a while ago. Anyways it seems a Netlify form would be nice for the contact us page, and the job postings page would be nice as a blog template. Do you want me to add this and submit a pull request? Or what changes do you want to be made?